+++ connection_helper.go
@@ -14,7 +14,7 @@ import (
 func (c *Connection) writeData(data []byte) error {
 	_, err := c.conn.Write(data[:])
 	if err != nil {
-		return ErrRQLConnection{err.Error()}
+		return RQLConnectionError{err.Error()}
 	}
 
 	return nil
@@ -26,7 +26,7 @@ func (c *Connection) writeHandshakeReq() error {
 
 	data := c.buf.takeSmallBuffer(dataLen)
 	if data == nil {
-		return ErrRQLDriver{ErrBusyBuffer.Error()}
+		return RQLDriverError{ErrBusyBuffer.Error()}
 	}
 
 	// Send the protocol version to the server as a 4-byte little-endian-encoded integer
@@ -56,14 +56,14 @@ func (c *Connection) readHandshakeSuccess() error {
 		if err == io.EOF {
 			return fmt.Errorf("Unexpected EOF: %s", string(line))
 		}
-		return ErrRQLConnection{err.Error()}
+		return RQLConnectionError{err.Error()}
 	}
 	// convert to string and remove trailing NUL byte
 	response := string(line[:len(line)-1])
 	if response != "SUCCESS" {
 		response = strings.TrimSpace(response)
 		// we failed authorization or something else terrible happened
-		return ErrRQLDriver{fmt.Sprintf("Server dropped connection with message: \"%s\"", response)}
+		return RQLDriverError{fmt.Sprintf("Server dropped connection with message: \"%s\"", response)}
 	}
 
 	return nil
@@ -75,7 +75,7 @@ func (c *Connection) writeQuery(token int64, q []byte) error {
 
 	data := c.buf.takeBuffer(dataLen)
 	if data == nil {
-		return ErrRQLDriver{ErrBusyBuffer.Error()}
+		return RQLDriverError{ErrBusyBuffer.Error()}
 	}
 
 	// Send the protocol version to the server as a 4-byte little-endian-encoded integer