+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+In an attempt to make this library more "idiomatic" some functions have been renamed, for the full list of changes see below.
+
+### Changed 
+ - Renamed `Db` to `DB`
+ - Renamed `DbCreate` to `DBCreate`
+ - Renamed `DbDrop` to `DBDrop`
+ - Renamed `RqlConnectionError` to `RQLConnectionError`
+ - Renamed `RqlDriverError` to `RQLDriverError`
+ - Renamed `RqlClientError` to `RQLClientError`
+ - Renamed `RqlRuntimeError` to `RQLRuntimeError`
+ - Renamed `RqlCompileError` to `RQLCompileError`
+
 ## v0.7.2 - 2015-05-05
 ### Added
  - Added support for connecting to a server using TLS (#179)